# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Universal Document Vector Search Service",
    "description": "Микросервис для векторного поиска по документам с поддержкой метаданных, категорий и версионирования",
    "version": "1.2.0"
  },
  "paths": {
...
    "/v1/categories/search": {
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Поиск по категориям",
        "description": "Выполняет семантический поиск по иерархии категорий.\n\n**Параметры запроса (в JSON body):**\n- `query_text` — текст запроса для поиска (обязательно)\n- `limit` — максимальное количество результатов (1–100, по умолчанию 10)\n- `grouped` — если `true`, возвращает категории, сгруппированные по коллекциям (по умолчанию `false`)\n- `fields` — список полей для возврата (опционально). Если не указан, возвращаются только `path` и `score`.\n\n**Доступные поля для `fields`:**\n- `path` — полный путь категории\n- `score` — релевантность категории\n- `id` — идентификатор категории\n- `category_name` — название категории\n- `categories` — список категорий в пути\n- `category_level` — уровень вложенности\n- `category_id` — идентификатор категории\n- `levels.category_level0`, `levels.category_level1`, ... — уровни категорий\n- `id_levels.category_id_level0`, `id_levels.category_id_level1`, ... — идентификаторы уровней\n\n**Формат запроса (минимальный):**\n```json\n{\n  \"query_text\": \"документация API\",\n  \"limit\": 10\n}\n```\n\n**Формат ответа (по умолчанию - только path и score):**\n```json\n{\n  \"results\": [\n    {\"path\": \"Документация / API\", \"score\": 0.95},\n    {\"path\": \"Документация / Настройка\", \"score\": 0.87}\n  ]\n}\n```\n\n**Формат запроса (с полями):**\n```json\n{\n  \"query_text\": \"документация API\",\n  \"limit\": 10,\n  \"fields\": [\"path\", \"score\", \"category_name\"]\n}\n```\n\n**Формат ответа (с полями):**\n```json\n{\n  \"results\": [\n    {\n      \"path\": \"Документация / API\",\n      \"score\": 0.95,\n      \"category_name\": \"API\"\n    }\n  ]\n}\n```\n\n**Группировка по коллекциям:**\nКогда параметр `grouped=true`, результат возвращается в формате:\n```json\n{\n  \"results\": {\n    \"documents\": [\n      {\"path\": \"Документация / API\", \"score\": 0.95}\n    ],\n    \"documents_archive\": [\n      {\"path\": \"Архив / Справочник\", \"score\": 0.75}\n    ]\n  }\n}\n```\n\n**Особенности:**\n- Использует векторное сопоставление для поиска релевантных категорий\n- Возвращает полный путь категории (`category_path`) и уровень вложенности\n- **По умолчанию возвращает только `path` и `score`** для оптимизации производительности\n- При необходимости можно получить полную информацию, указав поля в `fields`\n- Результаты можно сгруппировать по коллекциям с помощью параметра `grouped`\n\nИспользуется для улучшения поиска документов через RRF-ранжирование, учитывая релевантность категорий.",
        "operationId": "search_categories_endpoint_v1_categories_search_post",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "grouped",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Если True, возвращает категории, сгруппированные по коллекциям",
              "default": false,
              "title": "Grouped"
            },
            "description": "Если True, возвращает категории, сгруппированные по коллекциям"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategorySearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Получение всех категорий",
        "description": "Возвращает список всех категорий из базы данных.\n\n**Параметры:**\n- `limit` — максимальное количество результатов (1–1000, по умолчанию 100)\n- `parent_id` — фильтр по родительской категории (опционально)\n\n**Формат ответа:**\n- Массив объектов `CategoryResponse` с информацией о каждой категории.",
        "operationId": "get_all_categories_v1_categories_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Максимальное количество результатов (1-1000)",
              "examples": [
                50
              ],
              "default": 100,
              "title": "Limit"
            },
            "description": "Максимальное количество результатов (1-1000)"
          },
          {
            "name": "parent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по родительской категории",
              "examples": [
                "uuid-категории-родителя"
              ],
              "title": "Parent Id"
            },
            "description": "Фильтр по родительской категории"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategorySearchRequest": {
        "properties": {
          "query_text": {
            "type": "string",
            "title": "Query Text",
            "description": "Текст запроса для поиска по категориям",
            "examples": [
              "Документация по API"
            ]
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Максимальное количество результатов (1-100)",
            "default": 10,
            "examples": [
              10
            ]
          },
          "grouped": {
            "type": "boolean",
            "title": "Grouped",
            "description": "Если True, возвращает категории, сгруппированные по коллекциям",
            "default": false,
            "examples": [
              false
            ]
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields",
            "description": "Список полей для возврата. Если не указан, возвращаются только `path` и `score`. Доступные поля: `id`, `score`, `category_name`, `category_path`, `categories`, `category_level`, `category_id`, `levels`, `id_levels`.",
            "examples": [
              [
                "path",
                "score"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "query_text"
        ],
        "title": "CategorySearchRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}